fix clang -Wshorten-64-to-32
src/common.c:190:37: error: implicit conversion loses integer precision: 'long' to 'uint32_t' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
                return data.lastupdated-mktime(&d)+timezone;
src/common.c:255:8: error: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Werror,-Wshorten-64-to-32]
        len = strlen(s);
src/cfg.c:676:17: error: implicit conversion loses integer precision: 'long' to 'int32_t' (aka 'int') [-Werror,-Wshorten-64-to-32]
                *cset->loci = strtol(value, (char **)NULL, 0);
src/ibw.c:259:12: error: implicit conversion loses integer precision: 'long' to 'int32_t' (aka 'int') [-Werror,-Wshorten-64-to-32]
                ivalue = strtol(value, (char **)NULL, 0);
src/fs.c:50:8: error: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Werror,-Wshorten-64-to-32]
        len = strlen(tmp);
src/fs.c:96:8: error: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Werror,-Wshorten-64-to-32]
        len = strlen(dir);
src/image.c:165:47: error: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Werror,-Wshorten-64-to-32]
                gdImageString(ic->im, gdFontGetTiny(), width-(strlen(datestring)*gdFontGetTiny()->w+12), 9+ic->showedge, (unsigned char*)datestring, ic->cheaderdate);

+++ src/ibw.c
@@ -183,7 +183,7 @@ int ibwcfgread(FILE *fd)
 {
 	char cfgline[512], name[512], value[512];
 	int i, j, linelen, count = 0;
-	int32_t ivalue;
+	long ivalue;
 
 	/* start from value search from first line */
 	rewind(fd);
@@ -258,10 +258,10 @@ int ibwcfgread(FILE *fd)
 		/* add interface and limit to list if value is within limits */
 		ivalue = strtol(value, (char **)NULL, 0);
 		if (ivalue<0 || ivalue>BWMAX) {
-			snprintf(errorstring, 512, "Invalid value \"%d\" for MaxBW%s, ignoring parameter.", ivalue, name);
+			snprintf(errorstring, 512, "Invalid value \"%ld\" for MaxBW%s, ignoring parameter.", ivalue, name);
 			printe(PT_Config);
 		} else {
-			ibwadd(name, ivalue);
+			ibwadd(name, (uint32_t)ivalue);
 		}
 	}
